+++ packages/ts-components/jest.config.js
@@ -24,7 +24,8 @@ const buildConfig = dir => {
     },
     globals: {
       'ts-jest': {
-        tsConfig: './tsconfig.jest.json'
+        tsConfig: './tsconfig.jest.json',
+        isolatedModules: true
       }
     },
     collectCoverage: true,