+++ packages/topic/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.10.99](https://github.com/newsuk/times-components/compare/@times-components/topic@5.10.98...@times-components/topic@5.10.99) (2024-08-28)
+
+**Note:** Version bump only for package @times-components/topic
+
+
+
+
+
+## [5.10.98](https://github.com/newsuk/times-components/compare/@times-components/topic@5.10.97...@times-components/topic@5.10.98) (2024-08-28)
+
+**Note:** Version bump only for package @times-components/topic
+
+
+
+
+
+## [5.10.97](https://github.com/newsuk/times-components/compare/@times-components/topic@5.10.96...@times-components/topic@5.10.97) (2024-08-28)
+
+**Note:** Version bump only for package @times-components/topic
+
+
+
+
+
+## [5.10.96](https://github.com/newsuk/times-components/compare/@times-components/topic@5.10.95...@times-components/topic@5.10.96) (2024-08-23)
+
+**Note:** Version bump only for package @times-components/topic
+
+
+
+
+
+## [5.10.95](https://github.com/newsuk/times-components/compare/@times-components/topic@5.10.94...@times-components/topic@5.10.95) (2024-08-22)
+
+**Note:** Version bump only for package @times-components/topic
+
+
+
+
+
+## [5.10.94](https://github.com/newsuk/times-components/compare/@times-components/topic@5.10.93...@times-components/topic@5.10.94) (2024-08-14)
+
+**Note:** Version bump only for package @times-components/topic
+
+
+
+
+
 ## [5.10.93](https://github.com/newsuk/times-components/compare/@times-components/topic@5.10.92...@times-components/topic@5.10.93) (2024-08-08)
 
 **Note:** Version bump only for package @times-components/topic